Move ValueType methods to KonanSymbols
